completed caterng add item

+++ src/actions/delete-catering-menu-action.ts
@@ -1,14 +1,24 @@
 "use server";
 
-import { deleteFile, extractPublicId } from "@/config/cloudinary.config";
+import { deleteFile } from "@/config/cloudinary.config";
 import connectDB from "@/lib/mongodb";
+import { error400 } from "@/lib/response";
 import CateringMenu from "@/models/cateringMenuModel";
+import Catering from "@/models/cateringModel";
+import mongoose from "mongoose";
 import { revalidatePath } from "next/cache";
 
 export async function deleteCateringMenuAction(id: string) {
     try {
         await connectDB();
 
+        const objectId = mongoose.Types.ObjectId.createFromHexString(id);
+        const isMenuInUse = await Catering.findOne({
+            items: { $elemMatch: { itemId: objectId } },
+        });
+        if (isMenuInUse) {
+            return { error: "Failed to delete menu item. It is in use." };
+        }
         const deletedMenu = await CateringMenu.findByIdAndDelete({ _id: id });
 
         if (!deletedMenu) {
